requirements.txt
runpod>=1.7.0
transformers>=4.51.0
huggingface_hub>=0.23.0
accelerate>=0.30.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
requests>=2.31.0
# torch already in base image (2.8.0 + CUDA 12.8.1)
